roomservice: Add --force-sync flag
Needed to fix the 'cannot overwrite a local work tree' errors when using roomservice.
They are triggered while syncing a dependency that is already present on the tree and has different HEAD.
We have several use scenarios where this error is present such as major version bumps or our build server.

Change-Id: I7ba903d0d06dcc0fd0c01cc4466d8a7defcba757