releasetools: Switch to dash'ed logging
Most of the OTA packages seem to use a dash before the action being
carried out when printing logs. This is beneficial to quickly finding
the start of the specific log message when looking at the output on
the device which in turn makes the whole process seem to be overseeable
easier.

Change-Id: I5c186a0a558a58ee884cfed6f4c841996fa88132