Add back all the build properties
Most notable of all, the build display ID was missing. That's the
one that is visible in the Settings app as the Build number. This
commit adds back that and more by correctly moving a code block
to be executed earlier which should have been done in the parent
commit which unified .prop file generation logics but was accidentally
missed.

Change-Id: I66c1f8f735bbc4d36701d59135d39b547b069831