roomservice: Ensure the project's remote exists
When adding a new project, the remote has to exist in the manifest
as well. If you try adding a project that uses CAF or something
along those lines, then you end up with a local manifest that is
all messed up and fails to sync, which is not what you want.. at all.
With this commit, the script fails right as early as possible
without modifying the actual local manifest in case some remote
is not declared.

Change-Id: I08bbbb2744c8aa31688c327eaf4060f6794fdbd6