build: reintroduce and rework BOARD_CUSTOM_BOOTIMG_MK flag
CyanogenMod has it's own recovery, it's own signing methods and all this
needs to be compressed and inligned accordingly.
AOSP introduced a flag that makes booting possible from a recovery.img
as well as the new ninja stuff.
Bypassing and forcing the usage of the old method here and possibly
breaking those devices that use the recovery.img to boot is not a good
idea.
Also, pushing the custom ones to the final zip has no use for us.
We should purge 'echo -e $COLOR' due to new ninja/kati implementation
not dealing with colors properly.

References:
https://github.com/AOSPA/android_build/commit/d80bef2b9e7aa5f92ab017247442798f47736566

Change-Id: I4d8a65f26ba41c3d635b58c94e12edaa73e059a2
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>