build: dynamically fix ARM erratum cortex-a53-843419
Before this patch we were applying
the ARM erratum fix cortex-a53-843419
for all targets.
That isn't optimal because applying the
fix to the linker results in a performance
penalty which applies to the output binary.

The fix is only needed for cortex-a53
based CPU targets. For targets like
"kryo" or "generic", we don't assume
the fix is needed.

Additionally it is to note that
the so called "errata fixes" are
workarounds for hardware issues,
thus the performance penalty.

Try to partially match the implementation
in ART and apply the fix when the target cpu
(armv8) is "cortex-a53" or "default".
For all other target CPUs, don't apply
the fix.
This is different from ART, because
ART requires to whitelist all target
CPUs. Build system doesn't do that
and it is clearly not a good idea to
open another, similar whitelist for
CPUs which don't need the fix.

Change-Id: I002d5dd4f63631c63cb7e59489be80b1f6817504
Signed-off-by: Alex Naidis <alex.naidis@linux.com>