Update patch set 1
Patch Set 1: Code-Review-1

This needs testing.
Would be cool if anyone with an arm64 and non a53 device could test this.

A little help:
Apply this patch which is up for review here to build/ and also this http://www.hastebin.com/liqocofune.pl to frameworks/rs .
If it errors out on a non-a53 build, this patch didn't work as intended.

Please let me know whether you can complete a successfull build on such a target or not.

Patch-set: 1
Label: Code-Review=-1