build: Handle dynamic product device names
On some devices we may want not to set
PRODUCT_DEVICE which would indicate that
we want to set ro.product.device from an init extension.

So, don't set ro.product.device to TARGET_DEVICE if
PRODUCT_DEVICE is empty.
Since TARGET_DEVICE is used for other functions in
our system, we need to set that instead of PRODUCT_DEVICE
for this case.

Change-Id: I6fa7a30e763d9e5e89acf53b9809b78c70d73a3a
Signed-off-by: Alex Naidis <alex.naidis@linux.com>