Update patch set 1
Patch Set 1:

i'm not sure if that's the way to go here.
All major silicon vendors (Samsung, Qualcomm, etc) ship kernel sources which compile with -mandroid.
So if we want to compile with -mno-android, we'd have to modify the sources for each device for this (i.e. revert the change you linked).

Patch-set: 1