Update patch set 1
Patch Set 1:

> I think if we would want this, we wouldn't want it for all devices
 > most likely

If the image is read randomly, it should.
No device would have better random performance by enabling compression.

But I'm not sure about this.
It's a basic OS 101 thing that an executable is first loaded to memory before execution.

We would need to first see if this is the case with ART.

Patch-set: 1