build: introduce USE_DEXPREOPT shell variable for pre-opting
This allows the build environment to check for the existence of a shell
variable to control dex-preoptimization globally.

With this new condition, if the variable is exported as "false" in bashrc:

export USE_DEXPREOPT="false"

All pre-opting values will be turned off in this case. Avoiding maintainers
to set those flags in their device trees.

If the variable is not exported or exported as "true" in bashrc:

export USE_DEXPREOPT="true"

Pre-opting will be enabled with the recommended values as per AOSP rules.

Change-Id: Icaabe83e53d18bfb4f8e406f85170d174b8eac47
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>