Copy dex file to javalib.jar for prebuilt JAVA_LIBRARIES.
Copy dex file to javalib.jar for prebuilt JAVA_LIBRARIES.
Right now classes.jar is copied to javalib.jar for prebuilts.
It causes an issue for PRODUCT_BOOT_JARS because dex_bootjars
expects javalib.jar to have dex file.But it doesn't have
classes.dex in it.
Add LOCAL_DEX_FILE variable to define the jar with dex to be copied.

CRs-Fixed: 2329897
Change-Id: I2d62c5f4ec76b9635924c801617a5cd67812f295