Update patch set 3
Patch Set 3:

do we want to use if exists or should we let build fail if the vendor/pa makefile is not there?  it's more obvious it's broken if build system yells at you

also, this seems like an awful lot of places to call that makefile, should we just if out the aosp one here and call ours in vendor/pa once?

Patch-set: 3
CC: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>