Don't generate recovery patch for A/B systems.
An A/B system that has a recovery image should not generate the recovery patch.
This change gates the creation of the recovery patch such that we create it
only when the recovery image exists and only for non-A/B systems.

Test: Build A/B system, verify that recovery patch is not present.
Bug: 144104493
Change-Id: Ib570c21b73912c4e4b7a964b192abc15fde9a753
(cherry picked from commit 0598bea04bf7d1e37f846aa7f14f6d1728b83089)