Guard Compile time metadata information with LOCAL_ABI_CHECKER
With this change additional metadata information will only be
captured for modules which will have LOCAL_ABI_CHECKER set to
true. Trimming of clag data is not needed as we limited scope
to particular modules which can be handled in module specific
make files.
Additionally , resetting abi_checker and include variables to
avoid stale information in module_info.json

CRs-Fixed: 2580673
Change-Id: Iab0c10c160d67df398add88c9a54a593f78d7d63