Handle the absence of 'partition=' tag for older vendor target-files
Since the 'partition=' tag is recently enabled, older vendor
target-files will not have this feature, and so, when we combine newer
framework target-files and older vendor target-files,
the merge_target_files script cannot handle it and errors out.
This change adds handling for that. It relaxes the partition tag need
on the vendor side, and merges the vendor side entries as-is in
apkcerts and apexkeys.
In the areas that further read these merged apkcerts/apexkeys
(i.e in sign_target_files_apks), the code already supports having
'partition=' tag as optional

Change-Id: I76cef7702c3265e269736d6f09a68c3a7b564b2d