[android-11.0.0_r38] Source rbesetup.sh from any location inside of a repo directory.
The current rbesetup.sh has to be source from the root directory
of the checked out repo since it has a relative path to the sourced
build/envsetup.sh script. A function was written in rbesetup.sh
script to source the rbesetup.sh script anywhere inside of a
checked out repo. Added additional error checking to validate
the envsetup.sh is available and not to continue sourcing the script
if envsetup.sh script is not found.

Fixes: b/169438960
Test: Sourced the rbesetup script from
        * external directory
	* inside of a symlink
	* outside of repo (sourcing fails which is correct)
	* root repo
Change-Id: I2db66a444074adca4d4ab87b8786e9044a3646f8
Merged-In: I2db66a444074adca4d4ab87b8786e9044a3646f8