Update patch set 4
Patch Set 4: Verified+1

This one is pretty much self explanatory.
I did however adapt the original commit to make it work with the PA flavor of the ota_from_target_files.py so please do take a very good look at it. It works/builds here and fixes the dynamic_partitions_op_list issue.
Without this commit it would essentially delete the dynamic partition group and build it again, which of course is pretty bad when there are partitions in there like vendor and odm that we are not building

Patch-set: 4
Reviewer: Gerrit User 1000588 <1000588@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Verified=+1