[memtag] Disable memtag for AndroidMk host modules.
BUILD_HOST_EXECUTABLE modules are substantially deprecated, but some
partners are still using them for their bits with the workaround
provided in the product definition. This fixes a build error where
the host module doesn't have a linkable ELF note archive.

MTE is not intended for host modules, and it's fine for us to say
"host module using AndroidMk - no MTE for you" if this changes.

Test: Manually tested using a BUILD_HOST_EXECUTABLE module.
CRs-Fixed: 2962468
Change-Id: Ifedff39f2f03c08bfb644221d2ab1b88e635c8a3